functions imported from get_functions module
